requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0
matplotlib>=3.7.0
xlsxwriter>=3.1.0